+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Ignite UI for Angular Change Log
 
 All notable changes for each version of this project will be documented in this file.
+## 6.1.2
+- `igxCombo` improvements
+    - Remote Data Binding fixes - selection preserving and keyboard navigation.
+
+    For more detailed information see the [official igxCombo documentation](https://www.infragistics.com/products/ignite-ui-angular/angular/components/combo.html).
+
 ## 6.1.1
 - `igxTimePicker` changes
     - `onClose` event added.